Fixes #4447
refactor the order import from being a class_eval on Spree::Order from api to a dedicated importer in spree_core to be reused by other parts. update the year in the licence files, added licence file to spree_core as well since this was missing.

remove the hub part from the build.sh

move this decorator to the correct place again

add the api refactor to changelogs

remove the order_decorator from api and added the method in core order model