Fix duplicate address records being created during the checkout process.
The bug happens when a user fills out their address details in the address step, continues to the delivery step, goes back to the address step, and continues to the delivery step again. When the user continues the checkout process the addresses are created as new records in the database rather than being updated. This is because the id parameter isn't permitted by the controller, without which the existing records cannot be updated, resulting in new records being created for every change to the order addresses.