Fixes issue that caused admin users to get unassociated from an order if a user_id parameter was not provided.
With these changes, when a user_id is provided it will either associate the provided user_id to an order if the value is not nil, or disassociate the admin user from the order if the value is nil. When a user_id is not provided, the order will stay associated with the admin user.