Make best-promotion query more deterministic
Mostly for the benefit of the specs.

This query can become non-deterministic when the amounts and created_at
timestamps of two promotion adjustments are the equal.  I believe this is most
likely to happen in the specs.  Especially in MySQL since MySQL does not store
fractional seconds.  We also saw the same behavior in Postgres when running
specs on CircleCI.

Timecop helps repro this issue.

(cherry picked from commit ecd181bf836a5a7e6cd72fd6c2449b4619451402 in
spree/spree)