Revert ReimbursementItem model
This removes ReimbursementItem model and squishes it back into the ReturnItem model.
We split this model out of ReturnItem and we still think that's a good idea.  But we've run into some issues with it so we'll have to look at adding it back in later when we can work through all the details.

Fixes #5160