Extract current order params to a function.
The parameters used to find and create orders is duplicated in 3 places.
Extracting them to a function reduces duplication, and will allow store
maintainers to place conditions on what makes an order unique. An
example would be store restrictions like spree-multi-domain, that
currently have to use an ugly module_eval.

Fixes #5164