Remove some bad factory values
These numbers are autogenerated by the models so we don't need to set them in the factories.
On top of that, the numbers were being set 1) without their normal prefixes, 2) to the exact same value for every instance of the factory, 3) to the exact same values in two different factories.  These things caused errors in our app specs due to cross-model collisions that were a bit of a pain to track down.

Fixes #5437