Add Carton table & model
This is the first of several steps to extract the concept of a "shipped"
shipment *out* of the Shipment model into a separate model.

A "Carton" is meant to represent a collection of inventory units that have
actually been shipped. A Carton must have inventory units and may have an
optional external_number (e.g. a 3PL identifier) in addition to a tracking
number.

Note: The name "Carton" is a bit confusing since it represents and actual
*shipment* whereas "Shipment" will now just represent a "pending shipment" or
"billed shipment" or something like that. However, it's not practical at the moment to rename the "Shipment"
model and table.  Also we're hopeful that eventually the Shipment model itself
will be further broken down and possibly go away entirely.

Conflicts:
	core/app/models/spree/inventory_unit.rb
	core/app/models/spree/stock_location.rb