New admin stock management
Adds the ability to manage stock through a dedicated admin UI. The new
UI allows an admin to make updates without incurring a full page reload
and without having to select the variant to update in a separate form.
It also introduces a variant search search form that uses
Spree::Core::Search::Variant to filter the returned results.
The same UI is present in the product's stock management section but the
variant search results are scoped to the current product.
The Configuration admin tab was renamed to Settings to allow more visual space
for the newly added Stock tab.

Conflicts:
	api/app/controllers/spree/api/stock_items_controller.rb
	api/spec/controllers/spree/api/stock_items_controller_spec.rb
	backend/app/controllers/spree/admin/products_controller.rb
	backend/app/views/spree/admin/shared/_tabs.html.erb
	backend/config/routes.rb
	backend/spec/controllers/spree/admin/products_controller_spec.rb
	backend/spec/features/admin/configuration/mail_methods_spec.rb
	backend/spec/features/admin/homepage_spec.rb
	backend/spec/features/admin/products/stock_management_spec.rb
	core/config/locales/en.yml