Remove fancy setup default country spec
The reset_preferences in spec_helper should mean we don't need to bother
keeping this variable around.

I've added a modifies_config metadata so in the future we can key off of
that on what needs preferences reset in an around block, rather than
relying on it happening for every test.