Remove unreasonable reset
There is a reason that typographical elements come with margin and padding,
so that you don't have to pad out each one you use individually.

This sweeping change affects suprisingly little because the admin area
is mostly table-based. However, I suspect that there will now be redundant
additions of padding and or margin in other areas.