Allow the ability to suppress a shipment confirmation mailer.
Sometimes shipments need to be manually marked as shipped for a store
that usually automatically marks shipments as shipped. This can be due
to dropped communications with their fulfillment provider.

In these cases, marking a shipment as shipped manually is mainly for
data cleanup, and sending a shipment confirmation can create confusion.

Fix specs

Add comment

Add more specs

Conflicts:
	api/spec/controllers/spree/api/shipments_controller_spec.rb
	backend/app/assets/javascripts/spree/backend/shipments.js.erb
	backend/app/assets/stylesheets/spree/backend/sections/_orders.scss
	backend/spec/features/admin/orders/order_details_spec.rb
	backend/spec/features/admin/orders/shipments_spec.rb
	core/app/models/spree/shipment.rb