Update template config/initializers/spree.rb
Changes the default to use_static_preferences! for any new stores.

Removes the old leading comment explaining the gotchas of changing
preferences in the admin. As there are no longer any preferences
changeable from the admin.

Adds a small selection of useful preferences, which new sites are most
likely to use.