Fix stock_management_spec ordering failure
This spec sometimes failed when run in a random order. It was looking
for the field "#number-update-#{variant.id}", but the id used in that
css id was from stock_item, not variant. Fixed to use a class which
actually has the variant id.

Also remove a wait_for_ajax and redundant capybara statements.