Discriminate on the customer details display by permission
If the user can view an order, but not update it. Link to the show
action instead of the edit action. The customer details controller is a
little different than other admin controllers where edit can serve both
as a show action and an edit action due to the fact the edit action is
more simply just a form.