Replace broken Spree.t erb in JS file
This call would have always translated as the locale used to compile the
javascript. Instead the Spree.translations object should be used.

In the locale_spec, the locale is changed to french. If this spec was
the first one run which caused the JS assets to be compiled, they would
be compiled with a syntax error due to missing translations. Basically,
<%= Spree.t %> should never be used from JS, the Spree.translations
object should be used instead.