Store the user on the ability.
Permission sets need access to the user passed into the ability. It
makes sense to enforce our ability classes to respond to #user rather
than expanding the signature of permission sets to accept a user. As it
does not make sense to have a possibly different user from the ability
passed into the permission set.