Add a restricted transfer permission set.
This is an alternative to the StockManagement permission set. Instead
this conditionally authorizes specific actions for stock based on a
users associated stock locations.

The goal here, is to allow users who work in those locations, to manage
stock transfers and items for their locations, while still having read
only access to the rest of the interface.