Remove ContentController#show
Another not-a-vulnerability-but-looks-like-one. The render_404
before_filter guard was sufficient for preventing rendering of files
outside of views/content, views/store, views/base, and
views/application. The guard matches recommendations from the rails team
for preventing directory traversal in this situation.

However, this code is unused in every frontend store I've come across.
The behaviour this provides would be better done by a custom content
controller in the app itself, which would be trivial to do.