Only touch variants from option value if changed
Currently, we touch all matching variants on an option value save,
regardless of whether the option value has changed or not. However, that
touch can be very painful (if you save the size "L", that'll touch all
variants of that size which in our case is a lot). So it seems more
prudent to only do the touch if the option value actually changed.